pyproject.toml
[project]
name = "noyau-ai"
version = "0.1.0"
description = "NoyauAI - Daily tech digest for noyau.news"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "aiohttp>=3.11",
    "feedparser>=6.0",
    "openai>=1.58",
    "resend>=2.5",
    "loguru>=0.7",
    "jinja2>=3.1",
    "youtube-transcript-api>=0.6",
    "python-multipart>=0.0.17",
    "httpx>=0.28",
    "numpy>=2.0",
    "posthog>=3.7",
    # Video generation
    "moviepy>=2.0",
    "edge-tts>=6.1",
    "elevenlabs>=1.0",
    "google-auth>=2.36",
    "google-api-python-client>=2.150",
    "pillow>=11.0",
    # Storage
    "boto3>=1.35",
    # CLI
    "typer>=0.15",
    # Scheduling
    "apscheduler>=4.0.0a5",
]

[project.scripts]
noyau = "app.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "aiosqlite>=0.20",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
